podium-backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.5.0
django-storages==1.14.5
django-timezone-field==7.1
django_db==0.0.7
djangorestframework==3.15.2
filelock==3.17.0
filetype==1.2.0
fsspec==2025.2.0
ftfy==6.3.1
google-auth==2.38.0
google-genai==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
kombu==5.4.2
markdown2==2.5.3
markdownify==0.13.1
# marker-pdf==1.6.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
openai==1.65.1
opencv-python-headless==4.11.0.86
packaging==24.2
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
#psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMuPDF==1.25.3
PyMySQL==1.1.1
pypdfium2==4.30.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
surya-ocr==0.13.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
whitenoise==6.9.0
